add _boost and _parent field mapping support

+++ src/Nest/Domain/Mapping/ElasticProperty.cs
@@ -15,7 +15,7 @@ namespace Nest
 		public TermVectorOption TermVector { get; set; }
 		public FieldIndexOption Index { get; set; }
 
-		public decimal Boost { get; set; }
+		public double Boost { get; set; }
 
 
 		public string Analyzer { get; set; }
